Backend/app/features/categories/schemas.py
import uuid
from typing import List, Optional, Literal
from pydantic import BaseModel

CategoryType = Literal["EXPENSE", "INCOME", "INVESTMENT"]

class SubCategoryBase(BaseModel):
    name: str
    icon: Optional[str] = None
    color: Optional[str] = None
    type: CategoryType = "EXPENSE"
    is_surety: bool = False

class SubCategoryCreate(SubCategoryBase):
    category_id: uuid.UUID

class SubCategoryUpdate(BaseModel):
    name: Optional[str] = None
    icon: Optional[str] = None
    color: Optional[str] = None
    type: Optional[CategoryType] = None

class SubCategoryResponse(SubCategoryBase):
    id: uuid.UUID
    category_id: uuid.UUID
    user_id: Optional[uuid.UUID] = None
    is_surety: bool = False

    class Config:
        from_attributes = True

class CategoryBase(BaseModel):
    name: str
    icon: Optional[str] = None
    color: Optional[str] = None
    type: CategoryType = "EXPENSE"

class CategoryCreate(CategoryBase):
    pass

class CategoryUpdate(BaseModel):
    name: Optional[str] = None
    icon: Optional[str] = None
    color: Optional[str] = None
    type: Optional[CategoryType] = None

class CategoryResponse(CategoryBase):
    id: uuid.UUID
    user_id: Optional[uuid.UUID] = None
    sub_categories: List[SubCategoryResponse] = []

    class Config:
        from_attributes = True
