Backend/pyproject.toml
[project]
name = "Grip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "python-jose[cryptography]",
    "passlib[argon2]",
    "pydantic-settings",
    "email-validator",
    "httpx",
    "python-multipart",
    "psycopg2-binary",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "apscheduler>=3.11.2",
    "scipy>=1.17.0",
    "pandas>=2.3.3",
    "numpy>=2.4.1",
    "prophet>=1.2.1",
    "yfinance>=1.0",
]
